+++ .storybook/preview.tsx
@@ -1,12 +1,19 @@
 import NextImage from 'next/image';
+import { Open_Sans } from 'next/font/google';
 import { SiteProvider } from '../providers/siteProvider';
 import { ThemeProvider } from '../providers/themeProvider';
 import { LocaleProvider } from '../providers/localeProvider';
-import { openSans } from '../util/nextFonts';
 import type { Preview } from '@storybook/react';
 
 import '../styles/nodejs.dev/index.scss';
 import '../styles/new-design/index.scss';
+
+const openSans = Open_Sans({
+  weight: ['300', '400', '600', '700'],
+  display: 'fallback',
+  subsets: ['latin'],
+  variable: '--font-open-sans',
+});
 
 const preview: Preview = {
   parameters: {
